requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
urllib3
uritemplate
six
setuptools
rsa
requests
requests-oauthlib
pytz
pyparsing
pyasn1
pyasn1-modules
protobuf
packaging
oauthlib
idna
httplib2
googleapis-common-protos
google-auth
google-api-core
charset-normalizer
certifi
cachetools